refactor: 댓글 관련 로직 Comment 폴더 생성 및 파일 이동
